3.8/certificate/requirements.txt
opencv-python=4.4.0.44
numpy=1.19.2
Pillow=7.2.0
xlrd=1.2.0
configparser=5.0.1